@@ -7,8 +7,8 @@ versions.retrofitAdapterVersion = '2.2.0'
 versions.support = '27.1.+'
 versions.okhttpUrlConnection = '3.6.0'
 versions.okhttp_logging_interceptor = '3.12.+'
-versions.constraintlayout = '1.1.0'
-versions.glide = '3.7.0'
+versions.constraintlayout = '1.1.2'
+versions.glide = '4.7.1'
 versions.common3 = '3.5'
 versions.dagger = '2.10'
 versions.ormlite = '5.0'
@@ -16,7 +16,7 @@ versions.supportMultidexVersion = '1.0.2'
 versions.junit = '4.12'
 versions.espresso = '2.2.2'
 versions.guava = '21.0'
-versions.exoplayer = '2.6.1'
+versions.exoplayer = '2.8.2'
 ext.versions = versions
 def deps = [:]
 
@@ -68,9 +68,9 @@ ext.deps = deps
 
 def build_versions = [:]
 build_versions.min_sdk = 15
-build_versions.target_sdk = 27
-build_versions.build_tools = "27.0.3"
-build_versions.compileSdkVersion = 27
+build_versions.target_sdk = 28
+build_versions.build_tools = "28.0.1"
+build_versions.compileSdkVersion = 28
 ext.build_versions = build_versions
 
 def addRepos(RepositoryHandler handler) {
