@@ -1,7 +1,8 @@
 package me.sunzheng.mana;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 public class AccountManagerActivity extends AppCompatActivity {
 
