@@ -1,14 +1,14 @@
 package me.sunzheng.mana;
 
 import android.os.Bundle;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.AppCompatSpinner;
-import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.AdapterView;
 
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.appcompat.widget.AppCompatSpinner;
+import androidx.appcompat.widget.Toolbar;
 import me.sunzheng.mana.home.HomeApiService;
 import me.sunzheng.mana.home.mybangumi.FavoritesFragment;
 import me.sunzheng.mana.home.mybangumi.MyFavoritesPresenter;
