@@ -2,11 +2,11 @@ package me.sunzheng.mana.account;
 
 
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
+import androidx.fragment.app.Fragment;
 import me.sunzheng.mana.R;
 
 /**
