@@ -5,7 +5,6 @@ import android.content.Intent;
 import android.graphics.Color;
 import android.graphics.drawable.ColorDrawable;
 import android.os.Handler;
-import android.support.v7.widget.RecyclerView;
 import android.text.TextUtils;
 import android.view.LayoutInflater;
 import android.view.View;
@@ -14,11 +13,13 @@ import android.widget.ImageView;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 
-import com.bumptech.glide.DrawableTypeRequest;
 import com.bumptech.glide.Glide;
+import com.bumptech.glide.RequestBuilder;
+import com.bumptech.glide.request.RequestOptions;
 
 import java.util.List;
 
+import androidx.recyclerview.widget.RecyclerView;
 import me.sunzheng.mana.R;
 import me.sunzheng.mana.VideoPlayActivity;
 import me.sunzheng.mana.core.Episode;
@@ -67,13 +68,18 @@ public class EpisodeAdapter extends RecyclerView.Adapter<EpisodeAdapter.ViewHold
         } else {
             holder.itemView.setClickable(false);
         }
-        DrawableTypeRequest request = Glide.with(holder.itemView.getContext()).load(item.getThumbnail());
+        RequestBuilder request = Glide.with(holder.itemView.getContext()).load(item.getThumbnail());
+        RequestOptions options = null;
         if (item.getThumbnailImage() != null && !TextUtils.isEmpty(item.getThumbnailImage().dominantColor)
                 && item.getThumbnailImage().dominantColor.matches(RegexUtils.ColorPattern)) {
-            request.placeholder(new ColorDrawable(Color.parseColor(item.getThumbnailImage().dominantColor)));
+            options = new RequestOptions();
+            options.placeholder(new ColorDrawable(Color.parseColor(item.getThumbnailImage().dominantColor)));
         } else if (!TextUtils.isEmpty(item.getThumbnailColor()) && item.getThumbnailColor().matches(RegexUtils.ColorPattern)) {
-            request.placeholder(new ColorDrawable(Color.parseColor(item.getThumbnailColor())));
+            options = new RequestOptions();
+            options.placeholder(new ColorDrawable(Color.parseColor(item.getThumbnailColor())));
         }
+        if (options != null)
+            request.apply(options);
         request.into(holder.mImageView);
         holder.mEpisodeNoTextView.setText(holder.itemView.getContext().getString(R.string.episode_template, item.getEpisodeNo() + ""));
         holder.mTitleTextView.setText(LanguageSwitchUtils.switchLanguageToJa(holder.itemView.getContext(), item.getName(), item.getNameCn()));
