@@ -3,9 +3,9 @@ package me.sunzheng.mana.home.mybangumi;
 import android.app.Service;
 import android.content.Intent;
 import android.os.IBinder;
-import android.support.annotation.Nullable;
 import android.util.SparseArray;
 
+import androidx.annotation.Nullable;
 import me.sunzheng.mana.core.BangumiModel;
 
 /**
