@@ -2,8 +2,6 @@ package me.sunzheng.mana.widget;
 
 import android.content.Context;
 import android.os.Handler;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.support.v4.media.session.MediaControllerCompat;
 import android.util.AttributeSet;
 import android.view.SurfaceView;
@@ -12,6 +10,8 @@ import android.widget.FrameLayout;
 
 import com.google.android.exoplayer2.SimpleExoPlayer;
 
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 import me.sunzheng.mana.R;
 
 /**
